requirements.txt
redis
python-dotenv
pytest
